+++ tactics2d/envs/dataset_replay_base.py
@@ -1,0 +1,6 @@
+import gymnasium as gym
+
+
+class DatasetReplayEnvBase(gym.Env):
+    def __init__(self):
+        super().__init__()